pyproject.toml
[project]
name = "basemodel-agent"
version = "0.1.0"
description = "Restack-native BaseModel Agent with planning, memory, and tool execution"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "restack_ai",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
